Move buildMarkdownPage utility into llms-txt scripts
